[REFACTOR] Phase 2: derive seed mappings, share marketplace Data table (task 2.8)
- MARKETPLACE_SEED_RESOURCE_KEY_TO_ID derived from MARKETPLACE_RESOURCE
  (fixture keys are the lowercase constant names) instead of restating
  every resource
- Extract the marketplace Data pgTable into a client-free
  services/marketplace/src/schema.ts imported by both db.ts and
  scripts/seed.ts, removing the near-copy in seed.ts (the copies had
  already drifted on the jsonb $type)
- Document the seeded ServiceResource snake_case singular naming and its
  divergence from core.resources.json (nothing consumes either today)

Verified: bun run db:seed against dev DB, marketplace unit 12/12 and
e2e 4/4 against live api + marketplace.
